+++ main/geo/test/boofcv/alg/distort/pinhole/TestPinholePtoN2_F32.java
@@ -30,11 +30,11 @@ import static org.junit.Assert.assertEquals;
  * @author Peter Abeles
  */
 public class TestPinholePtoN2_F32 {
-	double fx = 200;
-	double fy = 300;
-	double skew = 1.2;
-	double x_c = 400;
-	double y_c = 450;
+	float fx = 200;
+	float fy = 300;
+	float skew = 1.2f;
+	float x_c = 400;
+	float y_c = 450;
 
 	@Test
 	public void stuff() {
@@ -53,6 +53,6 @@ public class TestPinholePtoN2_F32 {
 		GeometryMath_F32.mult(K_inv, in, expected);
 
 		assertEquals(expected.x,out.x,1e-5);
-		assertEquals(expected.y,out.y,1e-5);
+		assertEquals(expected.y, out.y, 1e-5);
 	}
 }
